% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_series_cuenta_financiera.R
\name{download_series_cuenta_financiera}
\alias{download_series_cuenta_financiera}
\title{Download cuenta financiera series from Banco de España.}
\usage{
download_series_cuenta_financiera(
  forcedownload = FALSE,
  forcegeneratecatalogcf = FALSE
)
}
\description{
This function downloads, stores and generates a catalog of the series contained in "Cuentas Financieras" dataset
}
\examples{
download_series_cuenta_financiera()
}
\keyword{banco}
\keyword{de}
\keyword{download}
\keyword{españa}
\keyword{full}
\keyword{series}
