% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_series_full.R
\name{download_series_full}
\alias{download_series_full}
\title{Download the full set of series Banco de España database}
\usage{
download_series_full(
  forcedownload = FALSE,
  forcegeneratecatalog = FALSE,
  forcecfdownload = FALSE,
  forcegeneratecatalogcf = FALSE,
  forceusebdecatalog = FALSE
)
}
\description{
This function downloads and stores the full catalog of Banco de España data series.
}
\examples{
download_series_full()
}
\keyword{banco}
\keyword{de}
\keyword{download}
\keyword{españa}
\keyword{full}
\keyword{series}
