% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_catalog.R
\name{generate_catalog}
\alias{generate_catalog}
\title{This function Generate a catalog of the series contained in the csv files located in a given directory passed as argument.}
\usage{
generate_catalog(directory, db = "")
}
\arguments{
\item{directory}{Character string. The name of the directory that contains the csv files to be processed.}

\item{db}{Character string. The name of the database that the series contained in the csv belong to.}
}
\description{
This function Generate a catalog of the series contained in the csv files located in a given directory passed as argument.
}
\keyword{banco}
\keyword{de}
\keyword{download}
\keyword{españa}
\keyword{full}
\keyword{series}
