% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_full_catalog.R
\name{generate_full_catalog}
\alias{generate_full_catalog}
\title{Generate full catalog from downloaded series.}
\usage{
generate_full_catalog()
}
\description{
This function generates a catalog from the series contained in all of Banco de España's datasets.
}
\examples{
generate_full_catalog()

}
\keyword{banco}
\keyword{catalog}
\keyword{de}
\keyword{españa}
\keyword{generate}
\keyword{series}
