% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_full_database.R
\name{generate_full_database}
\alias{generate_full_database}
\title{Generate a data object containing all the retrievable series in Banco de España series.}
\usage{
generate_full_database()
}
\description{
This function generates a feather and .Rda object containing all the series in Banco de España datasets.
}
\examples{
generate_full_database()

}
\keyword{Banco}
\keyword{BdE}
\keyword{España}
\keyword{data}
\keyword{database}
\keyword{de}
\keyword{full}
\keyword{generate}
\keyword{object}
\keyword{series}
