% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catalog.R
\name{get_catalog}
\alias{get_catalog}
\title{Return the full catalog of Banco de España series.}
\usage{
get_catalog()
}
\description{
Return the full catalog of Banco de España series.
}
\examples{
get_catalog()
}
\keyword{banco}
\keyword{catalog}
\keyword{de}
\keyword{dump}
\keyword{españa}
\keyword{full}
\keyword{series}
