% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_release_schedule.R
\name{get_release_schedule}
\alias{get_release_schedule}
\title{Estimate a release schedule for the series matching a certain string search.}
\usage{
get_release_schedule(search_str)
}
\arguments{
\item{get_release_schedule}{estimate release schedule}
}
\description{
This function takes a search string as argument and returns a dataframe containing the estimated date for the next data point.
}
\examples{
get_release_schedule()
}
\keyword{estimate}
\keyword{release}
\keyword{schedule}
