% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_series.R
\name{update_series}
\alias{update_series}
\title{This function downloads the datasets from Banco de España and replaces the existing CSV files containing the series (if any).}
\usage{
update_series()
}
\description{
This function downloads the datasets from Banco de España and replaces the existing CSV files containing the series (if any).
}
\examples{
update_series()
}
\keyword{banco}
\keyword{de}
\keyword{download}
\keyword{españa}
\keyword{full}
\keyword{series}
\keyword{update}
